+++ src/main/java/com/tms/service/NewsService.java
@@ -5,6 +5,8 @@ import com.tms.model.dto.NewsResponseDto;
 import com.tms.model.entity.News;
 import com.tms.repository.NewsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,10 +16,12 @@ import java.util.Optional;
 public class NewsService {
 
     private final NewsRepository newsRepository;
+    private final SecurityService securityService;
 
     @Autowired
-    public NewsService(NewsRepository newsRepository) {
+    public NewsService(NewsRepository newsRepository, SecurityService securityService) {
         this.newsRepository = newsRepository;
+        this.securityService = securityService;
     }
 
     public Optional<List<NewsResponseDto>> getAllNews() {
@@ -36,27 +40,36 @@ public class NewsService {
     }
 
     public Optional<NewsResponseDto> updateNews(News news) {
+        if(securityService.canAccessNews(news.getAuthorNewsId())) {
         return Optional.of(newsRepository.save(news))
                 .map(newsRequestDto -> NewsResponseDto.builder()
-                        .title(newsRequestDto.getTitle())
-                        .imageNews(newsRequestDto.getImageNews())
-                        .descriptionNews(newsRequestDto.getDescriptionNews())
-                        .authorNewsId(newsRequestDto.getAuthorNewsId())
-                        .build());
+                            .title(newsRequestDto.getTitle())
+                            .imageNews(newsRequestDto.getImageNews())
+                            .descriptionNews(newsRequestDto.getDescriptionNews())
+                            .authorNewsId(newsRequestDto.getAuthorNewsId())
+                            .build());
+        }
+        throw new AccessDeniedException("Access denied login:" + SecurityContextHolder.getContext().getAuthentication().getName() + " by id " + news.getAuthorNewsId());
     }
 
     public Boolean deleteNews(Long id) {
-        newsRepository.deleteById(id);
-        return !newsRepository.existsById(id);
+        if(securityService.canAccessNews(id)) {
+            newsRepository.deleteById(id);
+            return !newsRepository.existsById(id);
+        }
+        throw new AccessDeniedException("Access denied login:" + SecurityContextHolder.getContext().getAuthentication().getName() + " by id " + id);
     }
 
     public Optional<NewsResponseDto> createNews(News news) {
-        return Optional.of(newsRepository.save(news))
-                .map( newsRequestDto -> NewsResponseDto.builder()
-                        .title(newsRequestDto.getTitle())
-                        .imageNews(newsRequestDto.getImageNews())
-                        .descriptionNews(newsRequestDto.getDescriptionNews())
-                        .authorNewsId(newsRequestDto.getAuthorNewsId())
-                        .build());
+        if(securityService.canAccessNews(news.getAuthorNewsId())) {
+            return Optional.of(newsRepository.save(news))
+                    .map(newsRequestDto -> NewsResponseDto.builder()
+                            .title(newsRequestDto.getTitle())
+                            .imageNews(newsRequestDto.getImageNews())
+                            .descriptionNews(newsRequestDto.getDescriptionNews())
+                            .authorNewsId(newsRequestDto.getAuthorNewsId())
+                            .build());
+        }
+        throw new AccessDeniedException("Access denied login:" + SecurityContextHolder.getContext().getAuthentication().getName() + " by id " + news.getAuthorNewsId());
     }
 }
